First 2D map render

+++ Amberstar.Game/IPaletteIndexProvider.cs
@@ -4,10 +4,10 @@ namespace Amberstar.Game
 {
 	public interface IPaletteIndexProvider
 	{
-		int UIPaletteIndex { get; }
+		byte UIPaletteIndex { get; }
 
-		int GetTilesetPaletteIndex(int tileset);
-		int GetLabyrinthPaletteIndex(int paletteIndex); // not labyrinth index!
-		int Get80x80ImagePaletteIndex(Image80x80 image80X80);
+		byte GetTilesetPaletteIndex(int tileset);
+		byte GetLabyrinthPaletteIndex(int paletteIndex); // not labyrinth index!
+		byte Get80x80ImagePaletteIndex(Image80x80 image80X80);
 	}
 }
