+++ Amberstar.net/PaletteIndexProvider.cs
@@ -5,18 +5,18 @@ namespace Amberstar.net
 {
 	internal class PaletteIndexProvider
 	(
-		int uiPaletteIndex,
-		Dictionary<Image80x80, int> palettes80x80,
-		Dictionary<int, int> palettesTileset,
-		Dictionary<int, int> generalPalettes
+		byte uiPaletteIndex,
+		Dictionary<Image80x80, byte> palettes80x80,
+		Dictionary<int, byte> palettesTileset,
+		Dictionary<int, byte> generalPalettes
 	) : IPaletteIndexProvider
 	{
-		public int UIPaletteIndex => uiPaletteIndex;
+		public byte UIPaletteIndex => uiPaletteIndex;
 
-		public int Get80x80ImagePaletteIndex(Image80x80 image80X80) => palettes80x80[image80X80];
+		public byte Get80x80ImagePaletteIndex(Image80x80 image80X80) => palettes80x80[image80X80];
 
-		public int GetLabyrinthPaletteIndex(int paletteIndex) => generalPalettes[paletteIndex];
+		public byte GetLabyrinthPaletteIndex(int paletteIndex) => generalPalettes[paletteIndex];
 
-		public int GetTilesetPaletteIndex(int tileset) => palettesTileset[tileset];
+		public byte GetTilesetPaletteIndex(int tileset) => palettesTileset[tileset];
 	}
 }
